+++ lib/tasks/queue_faxes.rake
@@ -1,4 +1,4 @@
 desc "Queues faxes for signed, unfaxed applicants"
 task queue_faxes: :environment do
-  Export.enqueue_faxes
+  Enqueuer.new.enqueue_faxes
 end
